fix(dashboard): compact view formats states and offers a review path
Compact rows rendered raw enum values ("review_needed", "tv") where the
expanded cards show formatted labels, and an identity-review row's only
action was Cancel - no way to actually resolve the review from compact
view.

Extract CompactList into its own component (it was an unexported inline
in App.tsx, untestable), reuse the StateIndicator label map via a new
discStateLabel() helper so badges and list rows can never disagree,
widen the state column so "REVIEW NEEDED" isn't truncated, and add a
"Fix title" action for identity reviews mirroring the expanded card's
"Wrong title?" affordance.
