+++ frontend/src/app/components/StateIndicator.tsx
@@ -1,52 +1,19 @@
 import { motion } from "motion/react";
-import {
-  IcoIdle,
-  IcoScan,
-  IcoRipping,
-  IcoMatching,
-  IcoComplete,
-  IcoError,
-  IcoLibrary,
-  IcoReview,
-} from "./icons";
 import type { DiscState } from "./DiscCard";
-import { sv, type SvBadgeState } from "./synapse";
+import { sv } from "./synapse";
+import { DISC_STATE_CONFIG } from "./discState";
 
 interface StateIndicatorProps {
   state: DiscState;
 }
 
-interface StateConfig {
-  label: string;
-  badgeState: SvBadgeState;
-  color: string;
-  glow: string;
-  icon: React.ElementType;
-}
-
-/**
- * Map domain DiscState → Synapse v2 badge state + label + iconography.
- * Single source of truth for how each pipeline phase reads visually.
- */
-const STATE: Record<DiscState, StateConfig> = {
-  idle:           { label: "IDLE",          badgeState: "idle",     color: sv.inkDim,   glow: sv.inkDim,   icon: IcoIdle },
-  scanning:       { label: "SCANNING",      badgeState: "scanning", color: sv.yellow,   glow: sv.yellow,   icon: IcoScan },
-  review_needed:  { label: "REVIEW NEEDED", badgeState: "review",   color: sv.yellow,   glow: sv.yellow,   icon: IcoReview },
-  archiving_iso:  { label: "ARCHIVING",     badgeState: "matching", color: sv.purple,   glow: sv.purple,   icon: IcoLibrary },
-  ripping:        { label: "RIPPING",       badgeState: "ripping",  color: sv.magenta,  glow: sv.magenta,  icon: IcoRipping },
-  matching:       { label: "MATCHING",      badgeState: "matching", color: sv.amber,    glow: sv.amber,    icon: IcoMatching },
-  organizing:     { label: "ORGANIZING",    badgeState: "matching", color: sv.purple,   glow: sv.purple,   icon: IcoMatching },
-  processing:     { label: "PROCESSING",    badgeState: "matching", color: sv.amber,    glow: sv.amber,    icon: IcoMatching },
-  completed:      { label: "COMPLETE",      badgeState: "complete", color: sv.green,    glow: sv.green,    icon: IcoComplete },
-  error:          { label: "ERROR",         badgeState: "error",    color: sv.red,      glow: sv.red,      icon: IcoError },
-};
-
 /**
  * State pill — Synapse v2 Sv badge styling with a Lucide icon prefix.
  * Public prop contract is unchanged so existing callers keep working.
+ * Label/color/icon mapping lives in discState.ts (shared with CompactList).
  */
 export function StateIndicator({ state }: StateIndicatorProps) {
-  const config = STATE[state];
+  const config = DISC_STATE_CONFIG[state];
   const Icon = config.icon;
   const isActive = state !== "completed" && state !== "error" && state !== "idle";
   const shouldSpin = state === "ripping" || state === "scanning";
